Fix #147539: Ctrl middle mouse up/down on search button crashes
We should sanitize how search data ownership is handled in handling
code, see blender/blender!149300. But this isn't a trivial change, that
we rather don't do so short before a release.

For now workaround the issue by disabling the feature to
increment/decrement values using Ctrl+Mousewheel over text buttons in
the troublesome cases. This isn't an important feature and was only
introduced in 4.5 (see 9741ced8c0). It's still working for most text
buttons, just not the one where data ownership is a problem.