Fix: Dropping collection assets with packing applies instance offset
Steps to reproduce using
- Download the Human Base Meshes demo file (from
  https://www.blender.org/download/demo-files/)
- Extract and place the .blend in an asset library
- Open an asset browser with the asset library, disable collection
  instancing in the Import Settings popover
- Drag collections from the base meshes bundle into the 3D viewport

Instead of being placed under the cursor, the objects had an offset
applied that matches the instance offset of their collection. Usually we
substract that offset to address this. This was getting the offset from
the wrong collection though.

Since 8e3214fe05, we make objects and collections local when using
packing. Apparently this changes the ID's session UID, which was already
passed to the drop operator properties. So the drop operator would fail
to find the just imported collection, and would silently fall back
(unintentionally) to the first collection in the file (!). So further
code would use the wrong collection to correct the object transforms.