Fix: Crash rendering multiple frames from the command line
This fix addresses crash when --render-frame command line argument
is given multiple frames, like: --render-frame 130..133.

The issue was introduced by 48c507316513d32726014f3070e176c4f506177f.
The crash is caused by the RE_RenderAnim() clearing the callbacks,
leading to a crash on a second call of RE_RenderAnim().

Thanks to Sergey for investigating this.