UI: Area Movement Snapping Based on Window Size
Implements To Do #148851. Currently Area size snapping positions are
based on even portions of the unobstructed travel distance. This means
they are spaced differently if any areas are in the way of dragging
from window edge to edge. This makes it difficult to have editors in
different workspaces match sizes exactly. This PR spaces the snapping
positions based on the window dimensions instead.