UI: Area Movement Soft Snapping at Region Heights
When resizing editors they currently soft snap to minimums, which is
the height of a header. This PR adds more soft snapping points for some
editors. Animation editors snap to the time scrubbing region. And if
showing a footer will snap there too. Console editor snaps to a point
that shows a single input line.