Vulkan: Reduce Pipelines
Currently it is possible that identical pipelines can be stored the
pipeline pool. The root cause is that state could be set that is part of
the pipeline key, but not used by the pipeline.

This change fixes this by clearing the pipeline key to only have data
that is relavant to the pipeline.