Fix #149765: VSE: Editing Speed Control "Frame Number" doesn't update Preview
Issue only arose when editing the frame number for the current frame at
least once. On playhead scrub, the problem is resolved, indicating an
issue w/ intraframe cache. (Indeed, the effect was grabbing an old
preprocessed frame for the previous "frame number").

To fix, just notify the intra-frame cache of the new `target_frame`.