BLI: add Set.lookup_key_or_add_cb method
This is very similar to the corresponding method on Map. It allows creating
the value that is stored in the set only when it does not exist already.

This is especially useful when the key contains extra data that is not part
of the hash, or when a value on the stack is used for the first check and
an allocation is only done of the key does not exist already.