Fix: BLI: bad handling of intrusive set/map slots with custom equality function
Set and Map have a special optimization for types where the "empty" or "removed"
state of a hash table slot can be encoded in the value itself. In practice, this mainly
affects pointer keys where -1 and -2 are these special values. This often saves 8 bytes
per slot because no separate state variable with padding is required.

This works fine overall. However, it could easily lead to a crash with a custom equality
function. That's because the special value may be passed into that function which
only expects valid values.

The fix is to do some extra checks when a non-default equality method is used. There
are also new unit tests to cover this case.

Note that this does not affect VectorSet because it does not have (nor need) this
optimization.