UI: Project Icon
This icon is a variation of the desktop icon.
It is intended now as a placeholder, but it may work
as a final icon if we do not come up with something
better by then.

This icon is not used at the moment. It will be first
used by the project setup branch (project_awareness).