Fix: Uninitialized variable access in Cycles
Shadow terminator tricks might access uninitialized value of `wo`,
which happens in cases BSDF evaluation returns LABEL_NONE (i.e. when
the incident ray is in the lower hemisphere).

It is likely harmless from the render result perspective, as the
`eval` in those cases is 0 and, intuitively, multiplying 0 by a
possibly uninitialized value is still 0. Still good to solve this
issue to allow use of debugging tools like Valgrind.