Fix #148196: Realize instance attributes to curve domain
When realizing instance attributes to the final geometry output of the
realize instances node, currentlty the results are always placed on the
point domain. Using the curve domain is more efficient because it
reduces memory usage and the amount of work necessary later on. In the
case of this bug report, the "extra work" was mixing which overflowed
integer attribute range.

For backward compatibility, keep the old behavior with versioning and a
new node option. For new nodes, the "legacy behavior" option is off by
default though.