UI: Snapping and Improved Feedback for IMAGE_OT_view_zoom
While resizing in the Image Editor using ctrl-wheel drag or using the
zoom gizmo, this adds snapping (to even 10% increments) while ctrl is
held. Status bar indicates that snapping is available and the state of
it. The current value of the zoom is shown on the areas's header like
many other operators. The only change from current behavior is that
Ctrl-wheel drag has snapping on by default, but can be turned off by
releasing ctrl while dragging.