Fix: Grease Pencil operations use uninitialized memory for view positions
The `view_positions` array is computed in many Grease Pencil operations based
on the final output point selection. However, the values read from the array
may include points outside of the point selection in some cases, such as
smoothing or averaging.

Change the set of computed view positions to match the actually needed range
of point values based on the index mask used in each case.