add BKE_modifiers_add_at_end_if_possible utility
This adds a modifier at the end of the stack, but respects if there are
modifiers which are "pinned to last" (put the new modifier before then).
Also take into account that some modifiers can only be added after
"only-deforming" modifiers (but need to stay before e.g. "generating"
modifiers).
Shares logic with #object_modifier_check_move_after().

Also used in a couple of ore places where pinning was not respected.