Fix: Vertex Paint crashes after redoing dyntopo toggle
When redoing, entering Sculpt Mode from Vertex Paint does not handle
mode switching correctly. This results in a potentially invalid state
for the Paint BVH if dynamic topology is enabled. To avoid crashing,
check for the correct BVH type when starting a stroke.

This is a simple mitigation intended to be backported to 5.0 and 4.5
LTS

Initially reported in #150014