Refactor: Mesh: Use strings to store active UV map
It's generally simpler to interact with the mesh's active and default
UV map attributes if they are stored as strings on the mesh. This is how
the status is stored for color attributes, and it's already the format
used for blend files. This commit switches the way the status is stored
at runtime. Most of the work has happened already as part of #122398,
this just changes the last few places and simplifies propagation of the
active status where now possible.

Note that in BMesh the active/default status is also still referenced
with the CustomData functions. I made this design process to avoid
changing _all_ the places that use `BM_uv_map_offsets_get`. The mesh
name strings and the CustomData active status in BMesh are kept in
sync, which is possible because they're only changed in a few places.

---

I verified that the uses of the following functions are either related
to BMesh or only used in versioning:
- CustomData_get_active_layer_index
- CustomData_get_active_layer
- CustomData_get_active_layer_name
- CustomData_set_layer_active
- CustomData_set_layer_active_index
- CustomData_get_layer
- CustomData_get_for_write
- CustomData_free_layer_active
- CustomData_get_render_layer_index
- CustomData_get_render_layer
- CustomData_get_render_layer_name
- CustomData_set_layer_render
- CustomData_set_layer_render_index