Cleanup: Mesh: Separate code path for clearing expanded edit undo data
The current code uses a "create" argument to distinguish the code path
that just frees the data that's "expanded" from the compressed edit mode
storage after it's no longer needed, commenting "having this argument
avoids having 2x code paths". Actually, "2x code paths" is better than
forcing the same function to do extremely different things, and even
allows a net removal of 6 lines.

The point of this change is to make changes for #122398.