Fix #148969: VSE: Can't add image sequences for files that don't follow naming scheme
New behavior is consistent with the image editor, but since it removes
functionality that existed in the VSE before, it's technically a
regression.

To fix, add a new `image_import_type` enum with choices between "Auto
Detect" (new 5.0 feature), "Image Sequence" (old 4.5 behavior), and
"Individual Images" (also new 5.0 feature).

Ideally, it would be nice to offer the "force" image detection option
for the image editor too, but that would require a substantial
refactor/rewrite, so for now, just partially revert !143974.