Fix #150024: Prevent depsgraph eval while RNA writing is disallowed
Python scripts can evaluated the depsgraph in UI callbacks, which should
not be possible when the `rna_disallow_writes` flag is set. This is an
internal flag in pyrna, only used for property access and operators.
This leaves a wide range of BKE functions accessible that modify data
in dangerous ways. Depsgraph evaluation is one of the most common cases.

This change gives read access to the RNA write flag (same as
`pyrna_write_check`) and uses that to skip depsgraph evaluation in
`CTX_data_ensure_evaluated_depsgraph`.