Refactor: Fix ubsan warnings related to enum bitwise not
Since facb17b0e3be ubsan would complain about ~enum_value not being a valid
value for the enum. Now add an intermediate type that is accepted by the
& and &= enum operators to avoid the warning and maintain type safety.

This slightly improves safety as well as it is no longer possible to
accidently write a = ~b instead of a &= ~b.