Fix: VSE Showing Z Axis Transform Constraint
While dragging strips in the VSE the Status Bar indicates that you can
constrain the movement to the X, Y, or Z axis. But Z does apply here
and is excluded by the keymap poll function. However the way that we
collapse repeating items to a more compact form skips this. This PR
just adds a check to make sure that Z is included.