Fix #149843: Eevee Reference Spheres offset to the bottom and cropped
Happens when when tool header and/or the 3d view header regions are
flipped to the bottom.

Eevee uses the outcome of `ED_region_visible_rect` for its internal
visible rect (with positive values offsetting from the bottom).
The way the GLSL shader positions the spheres, we have to negate the
"anchor" y in the shader.