Fix: Nodes: Dynamic labels for Combine Color shader node
A new way to set node socket labels was introduced in 3cf52069bf. Each
socket can now declare a callback function to choose its label string
dynamically, instead of redeclaring all labels in an update function.

An oversight led to the removal of the update function for the
Combine Color shader node, without introducing label callbacks for
each of the socket. As a result, its sockets were stuck in RGB mode.

This commit adds the callbacks needed for the labels to change
dynamically. They are basically copied from the Separate Color shader
node.