3x3 adjoint matrix now calculates using copied values
Some calls of the 3x3 adjoint had R and M pointing to the same matrix. The assert doesn't catch this since M is a const float pointer, why R is just a float pointer. We simply mirror the ideas in the other adjoint calculations, and copy the values of M to floats first.