Fix: UI text for Blend file failures
"Blend document" does not make sense for multiple reasons. First,
"Blend" doesn't mean anything, it's just the contraction we use for our
file extension. Second, Blender files are not really "documents," at
least we don't refer to them that way. This commit just corrects the
places that wording was used.