Fix #150124: Retiming strips results in incorrect speeed
This issue happens, when strip is retimed, split in half and retimed again by
moving first or last strip key. The problem is caused by not realizing "fake"
keys upon selection, because key that is selected already existed. However, the
key at split point was not created yet.

Since for connected strips fake keys are always realized, even if there aren't
any, do the same thing for strip that is being selected.