Render: Try to stay on the same layer and pass when switching slots
The currently selected render layer and slot is stored as an index in the
iuser previously was untouched when switching render slots, so if the
new slot has a different list of layers and/or passes within the active
layer, the active later/pass would change.

This changes the logic to try and keep the same layer/pass (by name) selected,
so when e.g. switching from a slot with layers A+B+C (with C active) to a slot
with A+C, the iuser->layer is updated from 2 to 1.

Another example where this occurs would be flipping between a previous render
with active compositor and an ongoing render, since the ongoing render doesn't
have the Composite layer yet.

TODO:
Currently, this does not work when switching render slots via the side panel,
since that goes through the RNA setters and I'm not sure how to get the
appropriate context there to find the iuser.
Making it read-only and adding a function to set the slot would work I guess,
but seems excessive.