Fix #150538: ALT + F12 in default scene crashes
In 76c03744a809, we were already hiding the "Render Sequencer Image/
Animation" when certain conditions are not met (no sequencer scene and a
strip). One could still invoke/execute the operator via shortcuts
though, in this case blender could crash if there is no sequencer scene.

Now guard against that in `invoke` and `exec` as well.