Fix #150336: Vulkan Face Selection on AMD
Face selection uses face centers which rely on the vertex_size parameter
to be set. However in this case the vertex_size isn't set and could load
uninitialized memory. If the size is really large it will overwrite the
face geometry with face centers of other faces, resulting in selection
issues.

This PR fixes this by setting the vertex_size push constant for face
selection.