Fix #150414: Copying value to clipboard causes Steam to crash on Linux X11
The issue was caused by an oversight in 720fc44ca3 where atoms
previously defined directly in the functions where replaced by accesses
to the `atom_` struct to avoid unecessarily redeclaring them.

Since it wasn't already present, the C_STRING atom was added to the
struct.  However, its initialization was ommitted, leaving it default
initialized / undefined. This could then cause external application
(such as Steam) to fail when obtaining this atom as part of a TARGETS
request.

Fixed by properly initializing the atom in the SystemX11 constructor.