realize_on_domain can use Anisotropic filter
The CPU BLI_ewa_filter() has been modified to not assume the input is
clipped at the border, and to reduce the maximum number of pixels
it samples for each output pixel from 260,100 to 289.

I did some measurements of execution time.

It appears the CPU implementation is about 4x slower than Box, but only
about 1.3x slower than Bspline.

The GPU implementation is MUCH faster, approaching the speed of bilinear
texture sampling. However it is unclear if my measurement included the
time needed to compute the mipmap.