Vulkan: Add support for front face dynamic state
Based on the (flipped) scaling of an object the front facing can be
different. This lead to additional pipelines when you duplicated and
flipped the object using scaling.

This PR adds support to make the front face a dynamic part of the
pipeline resulting in less pipeline variations. This will reduce the
number of pipeline compilations in !150583