Cleanup: Grease Pencil: Reword warning when rendering without depth pass
The warning used to be
`Warning: To render Grease Pencil, enable Combined and Depth passes`.
Which is a bit missleading since Grease Pencil will still render, but it
won't correctly be occluded.

So the changed wording now is:
`Warning: To correctly render occluded Grease Pencil objects,
enable Combined and Depth passes.`