WIP Core: preserve 'stable pointers' data accross undo steps.
Typically, a lot of addresses will remain the same accross undo steps,
so keeping the same mapping helps reducing the cost of the 'stable
pointers' process.

This allows to reduce the 'writing IDProps' cost of the
`Slow selection in Blender 5.0.blend` blendfile from #150350 report by
roughly 50%.

But even with this new code, the overall cost remains about 5 times what
it was in 4.5 (before introduction of pointers stability code).

Note that keeping the same 'stable pointers' data (the mapping etc.)
means that it will keep growing over each undo step writing. This is
probably fine in most cases (with the above file, which generates over
1.3M 'stable pointer ids', it adds about 800 new entries on each undo
step), as most pointers will remain the same - but might also lead to
issues in some edge-cases...