Fix #150560: Handle Invalid WinTab Reported Extent
Some (faulty) WinTab (Windows) drivers can report that their surface
width or height is zero. Otherwise they mostly work okay. But we use
these values when we normalize the values so we can compare against the
cursor position as returned by the OS.  This case results in a divide
by zero. This PR just tests for zero extent and returns that the values
are not "trusted" and then fall back to using Win32 positioning.