Fix #150590: VSE: Drag and drop from file browser doesn't snap
The functionality has existed for a while but was broken with a
transition to `Vector` snap datatype in 789737c.

That being said, it's fairly obscure since it only ever activated on
ctrl, supposedly because "other drag and drop code works the same" but I
couldn't find these cases, so now it takes the current snap state into
account.