Refactor: UI: Avoid legacy code in UI job progress template
This method of setting a per-block callback and assigning "return values" to
indivdual buttons to trigger the right logic in the callback is ancient. We
don't use it anymore in newer code since we have more direct and type-safe ways
to do the same. Here, I'm passing a specific callback to each button.