Refactor: Simplify mesh edit mode array store storage
First, instead of storing a map from a CustomData layer type to an array
of all the layers with that type, store each layer's data individually
in a map also using the layer name as a key. Second, instead of making
the CustomData invalid by temporarily removing the layer data, just free
the whole struct, and add layers back with the CustomData API.

This work is part of #122398, with the goal of making it simple to
extend this code to compress attributes in `AttributeStorage` instead.