Compositor does not generate_mipmaps unless they are needed
It appears that generate_mipmaps takes about 1/2 the time that bilinear sample of the
image does. So avoiding this could speed things up by approximately 1/3.

This version generates any needed mipmaps each time the input image is used. GPU_texture
should have a flag to know if the mipmaps are already updated to avoid this, but I did
not implement this yet as I wanted this patch to be an unobtrusive as possible. I also
think GPU_texture should have some convienence api to set all the sampling flags and
generate the mipmap using a single call.