UI: IMAGE_OT_view_zoom Only Snap With Gizmo
With #149947 IMAGE_OT_view_zoom allows snapping. This works great when
called via the "zoom in/out" gizmo, but is not so great when using
Ctrl-MMB drag as this means snapping is on by default. This PR removes
the snapping from the latter so it works like it previously did.