Paint: Move stencil and clone storage from Mesh to paint settings
For paint mode masking and cloning, two separate UV map active status
is currently stored on the mesh UV maps. These are really painting
editor settings though, rather than being properties of the UV map.
After changes to the Blender file format months ago, these properties
were also not saved in files. The problem hasn't been reported in that
time, so clearly these options see little use, and/or are used
transiently, meaning storing the status permanently in meshes is not
important.

For the `AttributeStorage` changes (#122398), we need to find a
different place to store the selected clone and stencil UV maps.
This commit moves the storage to the place they are used: the image
editor paint settings, and the selections are stored as strings.

This is a tradeoff with some downsides. Switching between meshes
with different UV map names might be more annoying, and the names
won't automatically be invalidated as the mesh changes. On the other
hand, the design is simpler, less centralized, more scalable, and
aligns better with the more modern attribute design, where data is
referenced by name across Blender. There is a reduction in internal
complexity too.

An alternative design is to store these two selected UV maps as strings
on the mesh, similar to how the active and default status are stored.