Object: Remove poll methods for transform conversion methods
Removes the `poll` callbacks that use
`context.selected_editable_objects` to avoid potential expensive calls
during UI redraw. This commit moves them to the `exec` callback to
provide a general warning to the user.

Note that this does not change the behavior of the "Apply" menu
(Ctrl-A), as these elements to not appear to be greyed out currently.