Fix #150862: Weight Paint gradient context menu missing/error
Caused by 4434a30d4042

Same fix applied as in 63cee2681470 (but this time to the
`VIEW3D_PT_paint_weight_context_menu`)

NOTE: The actual "why" of 63cee2681470 is not totally clear to me (as
in: why is the override that we use
[`context.tool_settings.weight_paint.unified_paint_settings`] actually
different from
`UnifiedPaintPanel.paint_settings(context).unified_paint_settings`? The
later resolves to
`context.tool_settings.weight_paint.unified_paint_settings` as well, no?

NOTE: I wonder if the same has to be done to
`_draw_tool_settings_context_mode` as well (the `def PAINT_WEIGHT` case)