Paint: Improve cancelling stroke support
Prior to this commit, the `PAINT_STROKE_MODAL_CANCEL` modal entry was
only valid for Sculpt mode. This commit makes the following changes:

* Makes behavior explicit by adding a warning if the key is used in
  an unsupported mode (Vertex Paint, Weight Paint), or in special cases
  in an otherwise supported mode (Dyntopo)
* Adds an explicit parameter to indicate if a stroke is being cancelled,
  so that undo steps are not erroneously created
* Adds support for canceling texture / image paint strokes

Supporting this in more paint modes is a rather large task, as doing so
requires a better delta based integration (e.g. Sculpt Undo, Image
Undo) to avoid a deep copy of the underlying attribute being painted.

Resolves #140798