Fix #151021: Crash with randomize color pressure
In certain situations, it is possible for the curvemap used to not be
initialized before a stroke, leading to a crash. Prior to this commit,
the randomize color curvemaps were only initialized on blend read, which
is a rather weak design in general and prone to errors. Since we know
these curves can be used when painting colors, initialize them when
starting the stroke with the paint brush in Sculpt mode, or generically
for Vertex and Texture paint.