DNA: Cover makesdna with regression tests
Quite rudimentary for now as the main goal is to make code testable
and not necessarily cover all possible corner cases.

The basic idea is to create SDNA from a set of controlled headers
which are not part of the Blender's DNA, and run checks on the
structures, members, sizes, and offsets.

In order to achieve this a new command line argument is added to the
makesdna binary which can be used to explicitly provide list of files
within the basepath to generate SDNA from. It is then used to create
a separate SDNA files which a regression test is linked against.

The motivation behind this change is to ensure basics of the logic
are covered, and that it is easy to add tests for corner cases before
doing further development of makesdna. There are a few ideas for the
further development:
- Use offsetof() as the source of truth for the offsets, and remove
  manual offset calculation. It also feels that it will remove the
  need to manually add padding.
- Use sizeof() for C arrays: essentially allow use of preprocessor
  defines for array definitions.
- Allow using enum types for DNA members.
  Currently they are mentioned in a comment at a best.