Fix: Compilation failure with WITH_TESTS_SINGLE_BINARY=OFF
TEST_SRC which is provided to the blender_add_test_suite_lib is supposed
to contain sources of tests. In some cases it was including some of the
utility sources.

Other cases were relying on the library created by the test utility
which is not something one should rely on.