Fix: USD: Make object names more deterministic during import
It's possible for identical constructs in USD to result in non-identical
constructs in Blender. For example, given a USD prim hierarchy at path
`/root/A/A` and an identical hierarchy at path `/root/B/B`, this can
result in either an object hierarchy of `A/A.001` or `B.001/B` in
Blender (or vice versa).

This is mostly due to our sorting step which attempts to work around
inefficencies when creating large amounts of objects; creating them in
order by name is faster. Since the incoming names are the same, e.g `A`
and `A`, sometimes the order is swapped.

To make things consistent now, and to alleviate problems in the future
with concurrent import, it's important to keep the order stable. We
break the tie on name by using the entire original USD path instead.