Fix #151082: Crashing upon adding bone under certain conditions
Crash happened when [1] no bone collection is active and [2] other bone
collection(s) were soloed.

Then, `ANIM_bonecoll_is_visible_editbone` would (rightfully) return
false but following code would de-reference non existing `bone-
>bone_collections.first`.

To resolve, catch the specific case and report this to the user (the
fact that the new bone wasnt added to a collection but is hidden because
other collection(s) are soloed).

Should still be backported to 5.0.1