Refactor: UI: Rename uiBut to blender::ui::Button, keeping alias
This has the benefit of sounding less weird, and not being much longer
once the "ui" prefix is removed in interface code which will be in the
blender::ui namespace. We discussed renaming uiBut to other words
like Widget, View, Control, Elem, or Component, which may be more
standard and imply less "pushing," but I left that discussion for later.