Assets: Disable editing catalogs in essentials asset library
Allowing this doesn't really make sense. Essentials assets cannot be
assigned to different catalogs anyway (without opening the source .blend
files, which users shouldn't do and will be reset in new Blender
versions). The fact that this was possible before was probably an
oversight, I don't think this was ever a conscious decision.